done
ja som to mal len som to necommitol :DDD

+++ task_03_taxi_fare/main.cpp
@@ -4,7 +4,7 @@
 #include <iostream>
 using namespace std;
 int main() {
-    long long km, suma = 0;
+    float km, suma = 0;
 
     cout << " zadaj kilometre " << endl;
     cin >> km;
